requirements.txt
fastapi>=0.115.0
google-genai>=0.3.0
openai>=1.0.0
pillow>=10.4.0
PyJWT
python-dotenv>=1.0.1
requests>=2.32.5
uvicorn[standard]>=0.30.0